# docs/guides/contract/editor-service-openapi.yaml
openapi: 3.0.3
info:
  title: Editor Service API
  version: 1.0.0
  description: |
    editor-service가 현재 제공하는 HTTP API 문서입니다.
    문서 CRUD, editor operation, block attachment, document snapshot, admin block 보조 API를 하나의 OpenAPI 문서로 정리합니다.
servers:
  - url: http://documents-app:8083
...
paths:
  /documents:
    get:
      tags: [Document]
      summary: Get my documents
      operationId: getDocuments
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Documents fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Document]
      summary: Create document
      operationId: createDocument
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponseEnvelope'
        '400':
          $ref: '#/components/responses/InvalidRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/DocumentNotFoundError'
...
  /documents/{documentId}/snapshots:
    post:
      tags: [DocumentSnapshot]
      summary: Create document snapshot
      operationId: createSnapshot
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '201':
          description: Snapshot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSnapshotResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/DocumentNotFoundError'
  /documents/{documentId}/snapshots/{snapshotId}:
    get:
      tags: [DocumentSnapshot]
      summary: Get document snapshot metadata
      operationId: getSnapshot
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/SnapshotId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Snapshot metadata fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSnapshotResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/DocumentSnapshotNotFoundError'
    delete:
      tags: [DocumentSnapshot]
      summary: Delete document snapshot
      operationId: deleteSnapshot
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/SnapshotId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Snapshot deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoidResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/DocumentSnapshotNotFoundError'
  /documents/{documentId}/snapshots/{snapshotId}/content:
    get:
      tags: [DocumentSnapshot]
      summary: Download document snapshot
      operationId: downloadSnapshot
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - $ref: '#/components/parameters/SnapshotId'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Snapshot content
          content:
            application/json:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/DocumentSnapshotNotFoundError'
...
components:
  parameters:
    UserId:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: string
      description: Gateway 또는 platform security bridge가 주입한 사용자 ID입니다.
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: 요청 추적 ID입니다.
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema:
        type: string
  responses:
    UnauthorizedError:
      description: Missing authentication context
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
    InvalidRequestError:
      description: Invalid request or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
...
    DocumentNotFoundError:
      description: Document not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
...
    DocumentSnapshotNotFoundError:
      description: Document snapshot not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
    DocumentVisibility:
      type: string
      enum:
        - PRIVATE
        - PUBLIC
...
    CreateDocumentRequest:
      type: object
      required:
        - title
      properties:
        parentId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          maxLength: 255
          example: 프로젝트 개요
        icon:
          $ref: '#/components/schemas/JsonObject'
        cover:
          $ref: '#/components/schemas/JsonObject'
...
    BaseResponseFields:
      type: object
      properties:
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponseFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            parentId:
              type: string
              format: uuid
              nullable: true
            title:
              type: string
            icon:
              $ref: '#/components/schemas/JsonObject'
            cover:
              $ref: '#/components/schemas/JsonObject'
            visibility:
              $ref: '#/components/schemas/DocumentVisibility'
            sortKey:
              type: string
              nullable: true
            createdBy:
              type: string
              nullable: true
            updatedBy:
              type: string
              nullable: true
            deletedAt:
              type: string
              format: date-time
              nullable: true
...
    DocumentSnapshotResponse:
      type: object
      properties:
        snapshotId:
          type: string
        documentId:
          type: string
          format: uuid
        documentVersion:
          type: integer
          format: int64
        originalName:
          type: string
        contentType:
          type: string
        size:
          type: integer
          format: int64
        status:
          type: string
          example: ACTIVE
        createdAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    VoidResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: string
          example: OK
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 요청 응답 성공
        code:
          type: integer
          example: 200
        data:
          nullable: true
    DocumentResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: string
          example: OK
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 요청 응답 성공
        code:
          type: integer
          example: 200
        data:
          $ref: '#/components/schemas/DocumentResponse'
    DocumentListResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: string
        success:
          type: boolean
        message:
          type: string
        code:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
...
    DocumentSnapshotResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: string
        success:
          type: boolean
        message:
          type: string
        code:
          type: integer
        data:
          $ref: '#/components/schemas/DocumentSnapshotResponse'
...
    ErrorResponseEnvelope:
      type: object
      properties:
        httpStatus:
          type: string
          example: FORBIDDEN
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 요청을 수행할 권한이 없습니다.
        code:
          type: integer
          example: 9003
        data:
          nullable: true